requirements.txt
beautifulsoup4==4.13.4
colorama==0.4.6
googlesearch_python==1.3.0
instaloader==4.14.1
requests==2.32.4
